refactor: split admin page classes into their own files
