+++ porches/generic/admin/dt-generic-porch-landing-tab-home.php
@@ -1,128 +1,7 @@
 <?php
 if ( ! defined( 'ABSPATH' ) ) { exit; } // Exit if accessed directly
 
-/**
- * Class DT_Ramadan_Porch_Menu
- */
-class DT_Ramadan_Porch_Landing_Menu implements DT_Porch_Admin_Menu_Interface {
-
-    public $token = 'dt_porch_generic';
-    public $title = 'Settings';
-
-    public function __construct() {
-        if ( ! is_admin() ) {
-            return;
-        }
-
-        if ( isset( $_POST['install_campaign_nonce'], $_POST["download_csv"], $_POST['selected_campaign'] )
-            && wp_verify_nonce( sanitize_text_field( wp_unslash( $_POST['install_campaign_nonce'] ) ), 'install_campaign_nonce' )
-        ){
-            add_action( "init", function (){
-                $campaign_id = sanitize_text_field( wp_unslash( $_POST['selected_campaign'] ) ); //phpcs:ignore
-                $this->download_campaign_csv( $campaign_id );
-                exit;
-            });
-        } else {
-            add_action( "admin_menu", array( $this, "register_menu" ) );
-            add_action( 'admin_enqueue_scripts', [ $this, 'scripts' ] );
-        }
-
-
-    } // End __construct()
-
-    /**
-     * Loads the subnav page
-     * @since 0.1
-     */
-    public function register_menu() {
-        if ( current_user_can( 'wp_api_allowed_user' ) ) {
-            add_submenu_page( 'edit.php?post_type=' . PORCH_LANDING_POST_TYPE, $this->title, $this->title, "edit_" . PORCH_LANDING_POST_TYPE, $this->token, [ $this, 'content' ], 0 );
-        }
-    }
-
-    public function scripts(){
-        wp_enqueue_script( 'dt_ramadan_script', plugin_dir_url( __FILE__ ) . '/admin.js', [
-            'jquery',
-        ], filemtime( plugin_dir_path( __FILE__ ) . '/admin.js' ), true );
-    }
-
-    /**
-     * Menu stub. Replaced when Disciple.Tools Theme fully loads.
-     */
-    public function extensions_menu() {}
-
-    /**
-     * Builds porch tab links
-     * @since 0.1
-     */
-    public function tab_headers( string $link ) {
-
-        $tab = $this->get_tab();
-
-        ?>
-
-        <a href="<?php echo esc_attr( $link ) . 'home' ?>" class="nav-tab <?php echo esc_html( ( $tab == 'home' || !isset( $tab ) ) ? 'nav-tab-active' : '' ); ?>">Home Page</a>
-        <a href="<?php echo esc_attr( $link ) . 'content' ?>" class="nav-tab <?php echo esc_html( ( $tab == 'content' ) ? 'nav-tab-active' : '' ); ?>">Install Starter Content</a>
-
-        <?php
-    }
-
-    public function tab_content() {
-
-        $tab = $this->get_tab();
-
-        switch ( $tab ){
-            case "home":
-                $object = new P4_Ramadan_Porch_Landing_Tab_Home();
-                $object->content();
-                break;
-            case "content":
-                $object = new P4_Ramadan_Porch_Landing_Tab_Starter_Content();
-                $object->content();
-                break;
-            default:
-                break;
-        }
-    }
-
-    private function get_tab(): string {
-        if ( isset( $_GET["tab"] ) ) {
-            return sanitize_key( wp_unslash( $_GET["tab"] ) );
-        }
-
-        return "";
-    }
-
-    public function download_campaign_csv( $campaign_id, $filename = "subscribers.csv", $delimiter = "," ){
-        $subscribers = DT_Posts::list_posts( "subscriptions", [ "campaigns" => [ $campaign_id ], "status" => [ "active" ], 'limit' => 1000 ], false );
-        if ( is_wp_error( $subscribers ) ){
-            return;
-        }
-        header( 'Content-Type: application/csv' );
-        header( 'Content-Disposition: attachment; filename="' . $filename . '";' );
-        $f = fopen( 'php://output', 'w' );
-
-        $array = [
-            [ "Name", "Email" ],
-        ];
-        foreach ( $subscribers["posts"] as $sub ){
-            $a = [
-                $sub['name'],
-            ];
-            if ( isset( $sub['contact_email'][0]["value"] ) ){
-                $a[] = $sub['contact_email'][0]["value"];
-            } else {
-                $a[] = '';
-            }
-            $array[] = $a;
-        }
-        foreach ( $array as $line ){
-            fputcsv( $f, $line, $delimiter );
-        }
-    }
-}
-
-class P4_Ramadan_Porch_Landing_Tab_Home {
+class DT_Generic_Porch_Landing_Tab_Home {
     public function content() {
         ?>
         <style>
@@ -565,208 +444,3 @@ class P4_Ramadan_Porch_Landing_Tab_Home {
     }
 
 }
-
-
-/**
- * Class P4_Ramadan_Porch_Tab_Second
- */
-class P4_Ramadan_Porch_Landing_Tab_Starter_Content {
-    public function content() {
-
-        ?>
-        <div class="wrap">
-            <div id="poststuff">
-                <div id="post-body" class="metabox-holder columns-1">
-                    <div id="post-body-content">
-                        <!-- Main Column -->
-
-                        <?php
-                        $object = new P4_Ramadan_Porch_Landing_Tab_Home();
-                        $object->box_campaign();
-
-                        $fields = DT_Prayer_Campaigns::instance()->get_campaign();
-                        if ( ! empty( $fields ) ) {
-                            $this->main_column();
-                        }
-
-                        ?>
-
-                        <!-- End Main Column -->
-                    </div><!-- end post-body-content -->
-                </div><!-- post-body meta box container -->
-            </div><!--poststuff end -->
-        </div><!-- wrap end -->
-        <?php
-    }
-
-    public function main_column() {
-
-
-        global $wpdb;
-        if ( isset( $_POST['install_ramadan_nonce'] ) && wp_verify_nonce( sanitize_text_field( wp_unslash( $_POST['install_ramadan_nonce'] ) ), 'install_ramadan' ) ) {
-            $language = null;
-            if ( isset( $_POST["install_ramadan_language"] ) ){
-                $language = sanitize_text_field( wp_unslash( $_POST["install_ramadan_language"] ) );
-            }
-            $from_translation = null;
-            if ( isset( $_POST["install_ramadan_language_english"] ) ){
-                $language = sanitize_text_field( wp_unslash( $_POST["install_ramadan_language_english"] ) );
-                $from_translation = 'en_US';
-            }
-            if ( $language ){
-                $default_people_name = isset( $_POST[ $language . "_people_plural_masculine"] ) ? sanitize_text_field( wp_unslash( $_POST[ $language . "_people_plural_masculine"] ) ) : null;
-                $args = [
-                    "location_name" => isset( $_POST[ $language . "_location_name"] ) ? sanitize_text_field( wp_unslash( $_POST[ $language . "_location_name"] ) ) : null,
-                    "people_singular_masculine" => isset( $_POST[ $language . "_people_singular_masculine"] ) ? sanitize_text_field( wp_unslash( $_POST[ $language . "_people_singular_masculine"] ) ) : $default_people_name,
-                    "people_singular_feminine" => isset( $_POST[ $language . "_people_singular_feminine"] ) ? sanitize_text_field( wp_unslash( $_POST[ $language . "_people_singular_feminine"] ) ) : $default_people_name,
-                    "people_plural_masculine" => $default_people_name,
-                    "people_plural_feminine" => isset( $_POST[ $language . "_people_plural_feminine"] ) ? sanitize_text_field( wp_unslash( $_POST[ $language . "_people_plural_feminine"] ) ) : $default_people_name,
-                ];
-
-                P4_Ramadan_Porch_Starter_Content::load_content( $language, $args, $from_translation );
-            }
-
-            if ( isset( $_POST["delete_duplicates"] ) ){
-                $wpdb->query( "
-                    DELETE t1 FROM $wpdb->posts t1
-                    INNER JOIN $wpdb->posts t2
-                    LEFT JOIN $wpdb->postmeta t1m ON ( t1m.post_ID = t1.ID and t1m.meta_key = 'post_language')
-                    LEFT JOIN $wpdb->postmeta t2m ON ( t2m.post_ID = t2.ID and t2m.meta_key = 'post_language')
-                    WHERE t1.ID > t2.ID
-                    AND t1.post_title = t2.post_title
-                    AND ( t1m.meta_value = t2m.meta_value OR ( t1m.meta_value IS NULL AND t2m.meta_value IS NULL ) )
-                    AND t1.post_type = 'landing' AND t2.post_type = 'landing'
-                    AND ( t1.post_status = 'publish' OR t1.post_status = 'future' )
-                    AND ( t2.post_status = 'publish' OR t2.post_status = 'future' )
-                ");
-            }
-
-            if ( isset( $_POST["delete_posts"] ) ){
-                $language = sanitize_text_field( wp_unslash( $_POST["delete_posts"] ) );
-                $wpdb->query( $wpdb->prepare( "
-                    DELETE t1 FROM $wpdb->posts t1
-                    LEFT JOIN $wpdb->postmeta t1m ON ( t1m.post_ID = t1.ID and t1m.meta_key = 'post_language')
-                    WHERE t1m.meta_value = %s
-                    AND ( t1.post_status = 'publish' OR t1.post_status = 'future' )
-                    AND t1.post_type = 'landing'
-                ", $language ) );
-                if ( $language === "en_US" ){
-                    $wpdb->query( "
-                        DELETE t1 FROM $wpdb->posts t1
-                        LEFT JOIN $wpdb->postmeta t1m ON ( t1m.post_ID = t1.ID and t1m.meta_key = 'post_language')
-                        WHERE t1m.meta_value IS NULL
-                        AND ( t1.post_status = 'publish' OR t1.post_status = 'future' )
-                        AND t1.post_type = 'landing'
-                    " );
-                }
-            }
-        }
-        $languages = dt_ramadan_list_languages();
-        $fields = DT_Prayer_Campaigns::instance()->porch_fields();
-
-        $installed_langs_query = $wpdb->get_results( "
-            SELECT pm.meta_value, count(*) as count
-            FROM $wpdb->posts p
-            LEFT JOIN $wpdb->postmeta pm ON ( p.ID = pm.post_id AND meta_key = 'post_language' )
-            WHERE post_type = 'landing' and ( post_status = 'publish' or post_status = 'future')
-            GROUP BY meta_value
-        ", ARRAY_A);
-        $installed_langs = [];
-        foreach ( $installed_langs_query as $result ){
-            if ( $result["meta_value"] === null ){
-                $result["meta_value"] = 'en_US';
-            }
-            if ( !isset( $installed_langs[$result["meta_value"]] ) ){
-                $installed_langs[$result["meta_value"]] = 0;
-            }
-            $installed_langs[$result["meta_value"]] += $result["count"];
-        }
-
-        ?>
-        <form method="post">
-            <?php wp_nonce_field( 'install_ramadan', 'install_ramadan_nonce' ) ?>
-            <!-- Box -->
-            <table class="widefat striped">
-                <thead>
-                <tr>
-                    <th>Install Starter Ramadan Content</th>
-                </tr>
-                </thead>
-                <tbody>
-                    <tr><td>
-                            <p>See Prayer Fuel instructions <a href="https://github.com/Pray4Movement/pray4ramadan-porch/wiki/Prayer-fuel" target="_blank">here</a> </p>
-                    <table>
-                        <thead>
-                        <tr>
-                            <th>Language</th>
-                            <th>Installed Posts</th>
-                            <th>Location Name</th>
-<!--                            <th>People Singular Masculine</th>-->
-<!--                            <th>People Singular Feminine</th>-->
-                            <th>People Plural Masculine</th>
-                            <th>People Plural Feminine</th>
-                            <th>Install</th>
-                            <th>Install in English</th>
-                            <th>Delete Posts</th>
-                            <th>Duplicates</th>
-                        </tr>
-                        </thead>
-                        <tbody>
-
-                        <?php foreach ( $languages as $language_key => $language ) :
-                            $people_name = get_field_translation( $fields["people_name"], $language_key );
-                            $country_name = get_field_translation( $fields["country_name"], $language_key );
-                            $already_installed = ( $installed_langs[$language_key] ?? 0 ) > 0;
-                            $available_in_language = $language["prayer_fuel"] ?? false;
-                            $delete_enabled = ( $installed_langs[$language_key] ?? 0 ) > 1
-                            ?>
-                            <tr>
-                                <td><?php echo esc_html( $language["flag"] ); ?> <?php echo esc_html( $language['english_name'] ); ?></td>
-                                <td><?php echo esc_html( $installed_langs[$language_key] ?? 0 ); ?></td>
-                                <td><input style="width:150px" name="<?php echo esc_html( $language_key ); ?>_location_name" value="<?php echo esc_html( $country_name ); ?>"></td>
-<!--                                <td><input style="width:150px" name="--><?php //echo esc_html( $language_key ); ?><!--_people_singular_masculine" value="--><?php //echo esc_html( $people_name ); ?><!--"></td>-->
-<!--                                <td><input style="width:150px" name="--><?php //echo esc_html( $language_key ); ?><!--_people_singular_feminine" value="--><?php //echo esc_html( $people_name ); ?><!--"></td>-->
-                                <td><input style="width:150px" name="<?php echo esc_html( $language_key ); ?>_people_plural_masculine" value="" autocomplete="off"></td>
-                                <td><input style="width:150px" name="<?php echo esc_html( $language_key ); ?>_people_plural_feminine" value="" autocomplete="off"></td>
-                                <td>
-                                    <button type="submit" name="install_ramadan_language" class="button" value="<?php echo esc_html( $language_key ); ?>" <?php disabled( $already_installed || !$available_in_language ) ?>>
-                                        Install Ramadan Starter Content
-                                    </button>
-                                </td>
-                                <td>
-                                    <button type="submit" name="install_ramadan_language_english" class="button" value="<?php echo esc_html( $language_key ); ?>" <?php disabled( $already_installed ) ?>>
-                                        Install Content in English
-                                    </button>
-                                </td>
-                                <td>
-                                    <button type="button" onclick="jQuery('#confirm-delete-<?php echo esc_html( $language_key ); ?>').show()" class="button" <?php disabled( !$delete_enabled ) ?>>
-                                        Delete all <?php echo esc_html( $language["flag"] ); ?>
-                                    </button>
-                                    <div id="confirm-delete-<?php echo esc_html( $language_key ); ?>" style="display: none">
-                                        Are you sure? <button type="submit" class="button button-primary" name="delete_posts" value="<?php echo esc_html( $language_key ); ?>">Yes</button>
-                                        <button type="button" class="button button-secondary" onclick="jQuery('#confirm-delete-<?php echo esc_html( $language_key ); ?>').hide()">No</button>
-                                    </div>
-
-                                </td>
-                                <td>
-                                    <?php if ( ( $installed_langs[$language_key] ?? 0 ) > 60 ) :?>
-                                        <button type="submit" name="delete_duplicates" value="<?php echo esc_html( $language_key ); ?>">Delete Duplicates?</button>
-                                    <?php endif; ?>
-                                </td>
-                            </tr>
-
-                        <?php endforeach; ?>
-                        </tbody>
-                    </table>
-                    </td></tr>
-                </tbody>
-            </table>
-            <br>
-            <!-- End Box -->
-        </form>
-        <?php
-    }
-}
-
-
-
